# docs/swagger.yaml
openapi: 3.0.4
info:
  title: PPO Project.HttpServer
  version: v1
paths:
...
  '/api/v1/employees/{employeeId}':
    get:
      tags:
        - Employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
            text/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    delete:
      tags:
        - Employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    patch:
      tags:
        - Employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
            text/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /api/v1/employees:
    post:
      tags:
        - Employee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
            text/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
...
    CreateEmployeeDto:
      required:
        - birthday
        - duties
        - email
        - fullName
        - phoneNumber
        - photoPath
      type: object
      properties:
        fullName:
          type: string
          nullable: false
        phoneNumber:
          type: string
          nullable: false
        email:
          type: string
          nullable: false
        birthday:
          type: string
          format: date
          nullable: false
        photoPath:
          type: string
          nullable: true
        duties:
          type: string
          format: json
          nullable: true
      additionalProperties: false
...
    EmployeeDto:
      required:
        - birthday
        - email
        - employeeId
        - fullName
        - phoneNumber
        - photoPath
        - duties
      type: object
      properties:
        employeeId:
          type: string
          format: uuid
          nullable: false
        fullName:
          type: string
          nullable: false
        phoneNumber:
          type: string
          nullable: false
        email:
          type: string
          nullable: false
        birthday:
          type: string
          format: date
          nullable: false
        photoPath:
          type: string
          nullable: true
        duties:
          type: string
          nullable: true
      additionalProperties: false
    ErrorDto:
      required:
        - errorType
        - message
      type: object
      properties:
        errorType:
          type: string
          nullable: false
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    UpdateEmployeeDto:
      required:
        - fullName
        - phoneNumber
        - email
        - birthday
        - photoPath
        - duties
      type: object
      properties:
        fullName:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        birthday:
          type: string
          format: date
          nullable: true
        photoPath:
          type: string
          nullable: true
        duties:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
security:
  - Bearer: [ ]